refactor: Dynamic allocation of samplers to texture units
This is the first step of a critical refactor to how Iris manages samplers. Previously, samplers were allocated to a fixed texture unit, placing a hard cap on the number of total samplers that could be made available to shader programs.

Now, with the exception of a few reserved samplers managed externally (the primary texture used for textured content, the lightmap, and the overlay texture), Iris dynamically allocates samplers to texture units.

This means that a sampler only takes up a texture unit if it is used in the current program, opening the door to having a potentially unbounded number of samplers available, as long as each program only samples from a small selection of them.

Why now? This is how vanilla 1.17 manages samplers, so it makes things easier for Iris + Sodium on 1.17. It also allows us to finally move around the vanilla texture units back to how they are in vanilla. Finally, it allows the future complete unlocking of the limit of 8 color textures globally.

Overall, this refactor has been necessary for a long time, and now it has become critical. So I've decided to do it now, instead of piling on more kinda hacky code.