Reduce logo size to 7 kB with a 256x256 8-bit indexed color PNG
Previous versions ship a logo that is around 160kB 1280x1280 logo with 32-bit color. This adds substantially to a JAR file that is already over 1 megabyte. Reducing this file size will help slim down the JAR, though more work is needed.

Credit goes to RDKRACZ / KORR for spotting this and showing that an optimized resized icon can be way smaller than the original icon.

Closes https://github.com/IrisShaders/Iris/pull/678