Explicitly mark iris-explorer as incompatible
This mod was made for older versions of Iris which lacked
a shader pack selection screen. Now that Iris has its own
shader pack selection screen, it isn't needed, and the
author has explicitly marked the mod as deprecated.

I've decided to mark it as incompatible - while it still
might work, there's not really a reason to use it anymore.

It also references internal Iris code (in this case,
Iris.reload()), and relies on a given layout of the Iris
config file, which might change in the future. In the
interests of avoiding any future problems, it makes sense
to mark it as incompatible to ensure that nobody is using
it with new Iris versions.