Origins Enhanced for 1.16.5 is incompatible with Sodium 0.2.0 for 1.16.5
They do some hacky mixins into Sodium that only work with 0.1.0. Even though this doesn't affect Iris itself, it affects our Sodium fork.

Maybe this should be moved to the Sodium fork, but putting it here makes it more clear that Iris might be related, since only Sodium 0.1.x is available on CurseForge - and the only reason Sodium 0.2.x is loaded is because Iris is bundling it.