Finalize robustness improvements to shader skybox rendering
By moving our shader skybox rendering out of the renderSky method, we ensure it always
runs as expected, and isn't inadvertently cancelled by other mods that inject into
renderSky. The renderSky method is a wild west of mixins and patches, so we can't really
make many assumptions about what happens in that method.

In this case, FabricSkyboxes cancels it, and vanilla 1.18.2 tries to cancel it in high fog situations like the end.

When combine with 51dda9cb64112d0b62c93223edbf59aa93e5a931, this should address:
- https://github.com/IrisShaders/Iris/issues/1402 (We now only render the horizon with shaders enabled.)
- https://github.com/IrisShaders/Iris/issues/1224 (Fixed completely. You might need bulletproof-state-tracking on 1.16.x)
- https://github.com/IrisShaders/Iris/issues/131 (Addressed as best we can - FabricSkyboxes skyboxes are now assigned the CUSTOM_SKY render stage, which is the best we can do)