Support code to avoid a slow-path when copying depth textures
We need to make sure the copied textures have the same format as the main depth texture. Currently, while we avoid GL errors, we hit a massive driver slow-path on the format conversion.

We also use OpenGL 4.3 if it's available for a potentially faster way of copying depth texture content, through glCopyImageSubData. If not, we fall back to glBlitFramebuffer or glCopyTexSubImage2D as appropriate.