Remove incorrect vertex attribute input tracking code
Just because we don't have a tex/lightmap/overlay coordinate attribute in
the vertex data doesn't mean they can't be active - it's possible to use
glMultiTexCoord2f or similar. This was an incorrect assumption that was
derived from a different bug in DeferredWorldRenderingPipeline that caused
clouds and rain to be tinted red.