Enable GitHub actions artifact uploads
While in some cases we might not want builds of experimental branches to be uploaded, enabling artifact uploads allows people to easily test development versions of Iris. This is useful when bisecting an issue or just trying to see what a new commit does.

In the past we prevented this to avoid spurious issue reports. But it is clear now that it is more important to continuously be testing our development branches to ensure that bad bugs aren't popping up and not getting spotted until later.