fix: Work around a Sodium issue that caused corrupted block textures in some cases
This fixes a Sodium flaw that existed in its source code since February 2021 that became much easier to observe due to an Iris atlas stitching optimization released in June 2022. The bug was related to texture sprites allocated to the bottom and right edges of the texture atlas, which only occurs for a very small number of blocks. It also only affected large texture atlases, which were only likely with large mod packs or large combinations of resource packs, or water/lava when the water and lava textures somehow end up on the edge.

Though this is a Sodium flaw, it needs to be patched in Iris on Minecraft versions below 1.19 because Sodium only supports 1.19 and above.

Fixes:
- https://github.com/IrisShaders/Iris/issues/1678
- https://github.com/IrisShaders/Iris/issues/1502
- https://github.com/IrisShaders/Iris/issues/1726